feat: [console-lite] - top nav-bar - move v-logo to existing dir
