+++ libs/ui-toolkit/src/components/index.ts
@@ -29,5 +29,4 @@ export * from './text-area';
 export * from './theme-switcher';
 export * from './toggle';
 export * from './tooltip';
-export * from './v-logo';
 export * from './vega-logo';
