--- libs/ui-toolkit/src/components/v-logo/index.ts
@@ -1,1 +1,0 @@
-export * from './v-logo';
