--- libs/ui-toolkit/src/components/v-logo/v-logo.stories.tsx
@@ -1,11 +1,0 @@
-import type { Story, Meta } from '@storybook/react';
-import { VLogo } from './v-logo';
-
-export default {
-  component: VLogo,
-  title: 'V logo',
-} as Meta;
-
-const Template: Story = () => <VLogo />;
-
-export const Default = Template.bind({});
