--- libs/ui-toolkit/src/components/v-logo/v-logo.tsx
@@ -1,22 +1,0 @@
-import { t } from '@vegaprotocol/react-helpers';
-import React from 'react';
-
-export const VLogo = () => {
-  return (
-    <svg
-      aria-label={t('Vega logo')}
-      width="29"
-      height="34"
-      fill="currentColor"
-      viewBox="0 0 29 34"
-      className="mx-20 my-16"
-    >
-      <path d="M14.5003 29.1426H9.6665V34.0001H14.5003V29.1426Z" />
-      <path d="M19.3343 24.2859H14.5005V29.1434H19.3343V24.2859Z" />
-      <path d="M29.0008 19.4282H24.167V24.2857H29.0008V19.4282Z" />
-      <path d="M24.1673 0H19.3335V19.4283H24.1673V0Z" />
-      <path d="M9.66776 24.2859H4.83398V29.1434H9.66776V24.2859Z" />
-      <path d="M4.83378 0H0V24.2858H4.83378V0Z" />
-    </svg>
-  );
-};
