+++ libs/ui-toolkit/src/components/vega-logo/vega-logo.spec.tsx
@@ -1,10 +1,15 @@
 import { render } from '@testing-library/react';
 
-import { VegaLogo } from './vega-logo';
+import { VegaLogo, VLogo } from './vega-logo';
 
 describe('Vega logo', () => {
   it('should render successfully', () => {
     const { baseElement } = render(<VegaLogo />);
+    expect(baseElement).toBeTruthy();
+  });
+
+  it('V version should render successfully', () => {
+    const { baseElement } = render(<VLogo />);
     expect(baseElement).toBeTruthy();
   });
 });
