+++ libs/ui-toolkit/src/components/vega-logo/vega-logo.stories.tsx
@@ -1,5 +1,5 @@
 import type { Story, Meta } from '@storybook/react';
-import { VegaLogo } from './vega-logo';
+import { VegaLogo, VLogo } from './vega-logo';
 
 export default {
   component: VegaLogo,
@@ -7,5 +7,7 @@ export default {
 } as Meta;
 
 const Template: Story = () => <VegaLogo />;
+const TemplateVLogo: Story = () => <VLogo />;
 
 export const Default = Template.bind({});
+export const VVersion = TemplateVLogo.bind({});
